Round reported time to the `frequency` duration
The default frequency is `1m` so it will round down to the nearest
minute by default. If you override the default (e.g. `10s`) it will
round to that unit (second). It does not take into account the count so
it does not round to every 10 seconds if the frequency is 10s, for
example.